DoublyBufferedData allows bthread suspended
