+++ src/butil/containers/doubly_buffered_data.h
@@ -23,7 +23,6 @@
 #include <deque>
 #include <vector>                                       // std::vector
 #include <pthread.h>
-#include "butil/containers/doubly_buffered_data_internal.h"
 #include "butil/scoped_lock.h"
 #include "butil/thread_local.h"
 #include "butil/logging.h"
@@ -32,6 +31,7 @@
 #include "butil/errno.h"
 #include "butil/atomicops.h"
 #include "butil/unique_ptr.h"
+#include "butil/type_traits.h"
 
 namespace butil {
 
@@ -41,7 +41,8 @@ namespace butil {
 // modifications of data. As a side effect, this data structure can store
 // a thread-local data for user.
 //
-// Read(): begin with a thread-local mutex locked then read the foreground
+// --- `AllowSuspendBthread=false' ---
+// Read(): Begin with a thread-local mutex locked then read the foreground
 // instance which will not be changed before the mutex is unlocked. Since the
 // mutex is only locked by Modify() with an empty critical section, the
 // function is almost lock-free.
@@ -50,23 +51,54 @@ namespace butil {
 // foreground and background, lock thread-local mutexes one by one to make
 // sure all existing Read() finish and later Read() see new foreground,
 // then modify background(foreground before flip) again.
+//
+// But, when `AllowSuspendBthread=false', it is not allowed to suspend bthread
+// while reading. Otherwise, it may cause deadlock.
+//
+//
+// --- `AllowSuspendBthread=true' ---
+// It is allowed to suspend bthread while reading.
+// It is not allowed to use non-Void TLS.
+// If bthread will not be suspended while reading, it also makes Read() almost
+// lock-free by making Modify() *much* slower.
+// If bthread will be suspended while reading, there is competition among
+// bthreads using the same Wrapper.
+//
+// Read(): Begin with thread-local reference count of foreground instance
+// incremented by one which be protected by a thread-local mutex, then read
+// the foreground instance which will not be changed before its all thread-local
+// reference count become zero. At last, after the query completes, thread-local
+// reference count of foreground instance will be decremented by one, and if
+// it becomes zero, notifies Modify().
+//
+// Modify(): Modify background instance which is not used by any Read(), flip
+// foreground and background, lock thread-local mutexes one by one and wait
+// until thread-local reference counts which be protected by a thread-local
+// mutex become 0 to make sure all existing Read() finish and later Read()
+// see new foreground, then modify background(foreground before flip) again.
 
 class Void { };
 
-template <typename T, typename TLS = Void>
-class DoublyBufferedData {
-    friend class internal::WrapperTLSGroup<DoublyBufferedData>;
-    class Wrapper;
-public:
-    typedef T DataType;
+template <typename T> struct is_Void : false_type { };
+template <> struct is_Void<Void> : true_type { };
 
+template <typename T, typename TLS = Void, bool AllowBthreadSuspended = false>
+class DoublyBufferedData {
+    class Wrapper;
+    class WrapperTLSGroup;
+    typedef int WrapperTLSId;
+public:
     class ScopedPtr {
     friend class DoublyBufferedData;
     public:
-        ScopedPtr() : _data(NULL), _w(NULL) {}
+        ScopedPtr() : _data(NULL), _index(0), _w(NULL) {}
         ~ScopedPtr() {
             if (_w) {
-                _w->EndRead();
+                if (AllowBthreadSuspended) {
+                    _w->EndRead(_index);
+                } else {
+                    _w->EndRead();
+                }
             }
         }
         const T* get() const { return _data; }
@@ -77,6 +109,8 @@ public:
     private:
         DISALLOW_COPY_AND_ASSIGN(ScopedPtr);
         const T* _data;
+        // Index of foreground instance used by ScopedPtr.
+        int _index;
         Wrapper* _w;
     };
     
@@ -108,8 +142,73 @@ public:
     size_t ModifyWithForeground(Fn& fn, const Arg1&, const Arg2&);
     
 private:
-    const T* UnsafeRead() const
-    { return _data + _index.load(butil::memory_order_acquire); }
+    template <typename Fn>
+    struct WithFG0 {
+        WithFG0(Fn& fn, T* data) : _fn(fn), _data(data) { }
+        size_t operator()(T& bg) {
+            return _fn(bg, (const T&)_data[&bg == _data]);
+        }
+    private:
+        Fn& _fn;
+        T* _data;
+    };
+
+    template <typename Fn, typename Arg1>
+    struct WithFG1 {
+        WithFG1(Fn& fn, T* data, const Arg1& arg1)
+            : _fn(fn), _data(data), _arg1(arg1) {}
+        size_t operator()(T& bg) {
+            return _fn(bg, (const T&)_data[&bg == _data], _arg1);
+        }
+    private:
+        Fn& _fn;
+        T* _data;
+        const Arg1& _arg1;
+    };
+
+    template <typename Fn, typename Arg1, typename Arg2>
+    struct WithFG2 {
+        WithFG2(Fn& fn, T* data, const Arg1& arg1, const Arg2& arg2)
+            : _fn(fn), _data(data), _arg1(arg1), _arg2(arg2) {}
+        size_t operator()(T& bg) {
+            return _fn(bg, (const T&)_data[&bg == _data], _arg1, _arg2);
+        }
+    private:
+        Fn& _fn;
+        T* _data;
+        const Arg1& _arg1;
+        const Arg2& _arg2;
+    };
+
+    template <typename Fn, typename Arg1>
+    struct Closure1 {
+        Closure1(Fn& fn, const Arg1& arg1) : _fn(fn), _arg1(arg1) {}
+        size_t operator()(T& bg) { return _fn(bg, _arg1); }
+    private:
+        Fn& _fn;
+        const Arg1& _arg1;
+    };
+
+    template <typename Fn, typename Arg1, typename Arg2>
+    struct Closure2 {
+        Closure2(Fn& fn, const Arg1& arg1, const Arg2& arg2)
+            : _fn(fn), _arg1(arg1), _arg2(arg2) {}
+        size_t operator()(T& bg) { return _fn(bg, _arg1, _arg2); }
+    private:
+        Fn& _fn;
+        const Arg1& _arg1;
+        const Arg2& _arg2;
+    };
+
+    const T* UnsafeRead() const {
+        return _data + _index.load(butil::memory_order_acquire);
+    }
+
+    const T* UnsafeRead(int& index) const {
+        index = _index.load(butil::memory_order_acquire);
+        return _data + index;
+    }
+
     Wrapper* AddWrapper(Wrapper*);
     void RemoveWrapper(Wrapper*);
 
@@ -120,7 +219,7 @@ private:
     butil::atomic<int> _index;
 
     // Key to access thread-local wrappers.
-    internal::WrapperTLSId _wrapper_key;
+    WrapperTLSId _wrapper_key;
 
     // All thread-local instances.
     std::vector<Wrapper*> _wrappers;
@@ -146,20 +245,147 @@ template <typename T>
 class DoublyBufferedDataWrapperBase<T, Void> {
 };
 
-template <typename T, typename TLS>
-class DoublyBufferedData<T, TLS>::Wrapper
+// Use pthread_key store data limits by _SC_THREAD_KEYS_MAX.
+// WrapperTLSGroup can store Wrapper in thread local storage.
+// WrapperTLSGroup will destruct Wrapper data when thread exits,
+// other times only reset Wrapper inner structure.
+template <typename T, typename TLS, bool AllowSuspendBthread>
+class DoublyBufferedData<T, TLS, AllowSuspendBthread>::WrapperTLSGroup {
+public:
+    const static size_t RAW_BLOCK_SIZE = 4096;
+    const static size_t ELEMENTS_PER_BLOCK = (RAW_BLOCK_SIZE + sizeof(T) - 1) / sizeof(T);
+
+    struct BAIDU_CACHELINE_ALIGNMENT ThreadBlock {
+        inline DoublyBufferedData::Wrapper* at(size_t offset) {
+            return _data + offset;
+        };
+
+    private:
+        DoublyBufferedData::Wrapper _data[ELEMENTS_PER_BLOCK];
+    };
+
+    inline static WrapperTLSId key_create() {
+        BAIDU_SCOPED_LOCK(_s_mutex);
+        WrapperTLSId id = 0;
+        if (!_get_free_ids().empty()) {
+            id = _get_free_ids().back();
+            _get_free_ids().pop_back();
+        } else {
+            id = _s_id++;
+        }
+        return id;
+    }
+
+    inline static int key_delete(WrapperTLSId id) {
+        BAIDU_SCOPED_LOCK(_s_mutex);
+        if (id < 0 || id >= _s_id) {
+            errno = EINVAL;
+            return -1;
+        }
+        _get_free_ids().push_back(id);
+        return 0;
+    }
+
+    inline static DoublyBufferedData::Wrapper* get_or_create_tls_data(WrapperTLSId id) {
+        if (BAIDU_UNLIKELY(id < 0)) {
+            CHECK(false) << "Invalid id=" << id;
+            return NULL;
+        }
+        if (_s_tls_blocks == NULL) {
+            _s_tls_blocks = new (std::nothrow) std::vector<ThreadBlock*>;
+            if (BAIDU_UNLIKELY(_s_tls_blocks == NULL)) {
+                LOG(FATAL) << "Fail to create vector, " << berror();
+                return NULL;
+            }
+            butil::thread_atexit(_destroy_tls_blocks);
+        }
+        const size_t block_id = (size_t)id / ELEMENTS_PER_BLOCK;
+        if (block_id >= _s_tls_blocks->size()) {
+            // The 32ul avoid pointless small resizes.
+            _s_tls_blocks->resize(std::max(block_id + 1, 32ul));
+        }
+        ThreadBlock* tb = (*_s_tls_blocks)[block_id];
+        if (tb == NULL) {
+            ThreadBlock* new_block = new (std::nothrow) ThreadBlock;
+            if (BAIDU_UNLIKELY(new_block == NULL)) {
+                return NULL;
+            }
+            tb = new_block;
+            (*_s_tls_blocks)[block_id] = new_block;
+        }
+        return tb->at(id - block_id * ELEMENTS_PER_BLOCK);
+    }
+
+private:
+    static void _destroy_tls_blocks() {
+        if (!_s_tls_blocks) {
+            return;
+        }
+        for (size_t i = 0; i < _s_tls_blocks->size(); ++i) {
+            delete (*_s_tls_blocks)[i];
+        }
+        delete _s_tls_blocks;
+        _s_tls_blocks = NULL;
+    }
+
+    inline static std::deque<WrapperTLSId>& _get_free_ids() {
+        if (BAIDU_UNLIKELY(!_s_free_ids)) {
+            _s_free_ids = new (std::nothrow) std::deque<WrapperTLSId>();
+            if (!_s_free_ids) {
+                abort();
+            }
+        }
+        return *_s_free_ids;
+    }
+
+private:
+    static pthread_mutex_t _s_mutex;
+    static WrapperTLSId _s_id;
+    static std::deque<WrapperTLSId>* _s_free_ids;
+    static __thread std::vector<ThreadBlock*>* _s_tls_blocks;
+};
+
+template <typename T, typename TLS, bool AllowSuspendBthread>
+pthread_mutex_t DoublyBufferedData<T, TLS, AllowSuspendBthread>::WrapperTLSGroup::_s_mutex = PTHREAD_MUTEX_INITIALIZER;
+
+template <typename T, typename TLS, bool AllowSuspendBthread>
+std::deque<typename DoublyBufferedData<T, TLS, AllowSuspendBthread>::WrapperTLSId>*
+        DoublyBufferedData<T, TLS, AllowSuspendBthread>::WrapperTLSGroup::_s_free_ids = NULL;
+
+template <typename T, typename TLS, bool AllowSuspendBthread>
+typename DoublyBufferedData<T, TLS, AllowSuspendBthread>::WrapperTLSId
+        DoublyBufferedData<T, TLS, AllowSuspendBthread>::WrapperTLSGroup::_s_id = 0;
+
+template <typename T, typename TLS, bool AllowSuspendBthread>
+__thread std::vector<typename DoublyBufferedData<T, TLS, AllowSuspendBthread>::WrapperTLSGroup::ThreadBlock*>*
+        DoublyBufferedData<T, TLS, AllowSuspendBthread>::WrapperTLSGroup::_s_tls_blocks = NULL;
+
+template <typename T, typename TLS, bool AllowSuspendBthread>
+class DoublyBufferedData<T, TLS, AllowSuspendBthread>::Wrapper
     : public DoublyBufferedDataWrapperBase<T, TLS> {
 friend class DoublyBufferedData;
 public:
-    explicit Wrapper() : _control(NULL) {
+    explicit Wrapper()
+        : _control(NULL)
+        , _modify_wait(false) {
         pthread_mutex_init(&_mutex, NULL);
+        pthread_cond_init(&_cond[0], NULL);
+        pthread_cond_init(&_cond[1], NULL);
     }
     
     ~Wrapper() {
         if (_control != NULL) {
             _control->RemoveWrapper(this);
         }
+
+        if (AllowSuspendBthread) {
+            WaitReadDone(0);
+            WaitReadDone(1);
+        }
+
         pthread_mutex_destroy(&_mutex);
+        pthread_cond_destroy(&_cond[0]);
+        pthread_cond_destroy(&_cond[1]);
     }
 
     // _mutex will be locked by the calling pthread and DoublyBufferedData.
@@ -169,23 +395,77 @@ public:
         pthread_mutex_lock(&_mutex);
     }
 
+    // For `AllowSuspendBthread=true'.
+    inline void BeginReadRelease() {
+        pthread_mutex_unlock(&_mutex);
+    }
+
     inline void EndRead() {
         pthread_mutex_unlock(&_mutex);
+    }
+
+    // For `AllowSuspendBthread=true'.
+    // Thread-local reference count which be protected by _mutex
+    // will be decremented by one.
+    inline void EndRead(int index) {
+        BAIDU_SCOPED_LOCK(_mutex);
+        SubRef(index);
+        SignalReadCond(index);
     }
 
     inline void WaitReadDone() {
         BAIDU_SCOPED_LOCK(_mutex);
     }
+
+    // For `AllowSuspendBthread=true'.
+    // Wait until all read of foreground instance done.
+    inline void WaitReadDone(int index) {
+        BAIDU_SCOPED_LOCK(_mutex);
+        int& ref = index == 0 ? _ref[0] : _ref[1];
+        while (ref != 0) {
+            _modify_wait = true;
+            pthread_cond_wait(&_cond[index], &_mutex);
+        }
+        _modify_wait = false;
+    }
+
+    // For `AllowSuspendBthread=true'.
+    inline void SignalReadCond(int index) {
+        if (_ref[index] == 0 && _modify_wait) {
+            pthread_cond_signal(&_cond[index]);
+        }
+    }
+
+    // For `AllowSuspendBthread=true'.
+    void AddRef(int index) {
+        ++_ref[index];
+    }
+
+    // For `AllowSuspendBthread=true'.
+    void SubRef(int index) {
+        --_ref[index];
+    }
     
 private:
     DoublyBufferedData* _control;
     pthread_mutex_t _mutex{};
+    // For `AllowSuspendBthread=true'.
+    // _cond[0] for _ref[0], _cond[1] for _ref[1]
+    pthread_cond_t _cond[2]{};
+    // For `AllowSuspendBthread=true'.
+    // _ref[0] is reference count for _data[0],
+    // _ref[1] is reference count for _data[1].
+    int _ref[2]{0, 0};
+    // For `AllowSuspendBthread=true'.
+    // Whether there is a Modify() waiting for _ref0/_ref1.
+    bool _modify_wait;
 };
 
 // Called when thread initializes thread-local wrapper.
-template <typename T, typename TLS>
-typename DoublyBufferedData<T, TLS>::Wrapper* DoublyBufferedData<T, TLS>::AddWrapper(
-        typename DoublyBufferedData<T, TLS>::Wrapper* w) {
+template <typename T, typename TLS, bool AllowSuspendBthread>
+typename DoublyBufferedData<T, TLS, AllowSuspendBthread>::Wrapper*
+DoublyBufferedData<T, TLS, AllowSuspendBthread>::AddWrapper(
+        typename DoublyBufferedData<T, TLS, AllowSuspendBthread>::Wrapper* w) {
     if (NULL == w) {
         return NULL;
     }
@@ -207,9 +487,9 @@ typename DoublyBufferedData<T, TLS>::Wrapper* DoublyBufferedData<T, TLS>::AddWra
 }
 
 // Called when thread quits.
-template <typename T, typename TLS>
-void DoublyBufferedData<T, TLS>::RemoveWrapper(
-    typename DoublyBufferedData<T, TLS>::Wrapper* w) {
+template <typename T, typename TLS, bool AllowSuspendBthread>
+void DoublyBufferedData<T, TLS, AllowSuspendBthread>::RemoveWrapper(
+    typename DoublyBufferedData<T, TLS, AllowSuspendBthread>::Wrapper* w) {
     if (NULL == w) {
         return;
     }
@@ -223,14 +503,17 @@ void DoublyBufferedData<T, TLS>::RemoveWrapper(
     }
 }
 
-template <typename T, typename TLS>
-DoublyBufferedData<T, TLS>::DoublyBufferedData()
+template <typename T, typename TLS, bool AllowSuspendBthread>
+DoublyBufferedData<T, TLS, AllowSuspendBthread>::DoublyBufferedData()
     : _index(0)
     , _wrapper_key(0) {
+    static_assert(!(AllowSuspendBthread && !is_Void<TLS>::value),
+                  "Forbidden to allow suspend bthread with non-Void TLS");
+
     _wrappers.reserve(64);
     pthread_mutex_init(&_modify_mutex, NULL);
     pthread_mutex_init(&_wrappers_mutex, NULL);
-    _wrapper_key = internal::WrapperTLSGroup<DoublyBufferedData>::key_create();
+    _wrapper_key = WrapperTLSGroup::key_create();
     // Initialize _data for some POD types. This is essential for pointer
     // types because they should be Read() as NULL before any Modify().
     if (is_integral<T>::value || is_floating_point<T>::value ||
@@ -240,8 +523,8 @@ DoublyBufferedData<T, TLS>::DoublyBufferedData()
     }
 }
 
-template <typename T, typename TLS>
-DoublyBufferedData<T, TLS>::~DoublyBufferedData() {
+template <typename T, typename TLS, bool AllowSuspendBthread>
+DoublyBufferedData<T, TLS, AllowSuspendBthread>::~DoublyBufferedData() {
     // User is responsible for synchronizations between Read()/Modify() and
     // this function.
     
@@ -252,30 +535,43 @@ DoublyBufferedData<T, TLS>::~DoublyBufferedData() {
         }
         _wrappers.clear();
     }
-    internal::WrapperTLSGroup<DoublyBufferedData>::key_delete(_wrapper_key);
+    WrapperTLSGroup::key_delete(_wrapper_key);
     _wrapper_key = -1;
     pthread_mutex_destroy(&_modify_mutex);
     pthread_mutex_destroy(&_wrappers_mutex);
 }
 
-template <typename T, typename TLS>
-int DoublyBufferedData<T, TLS>::Read(
-    typename DoublyBufferedData<T, TLS>::ScopedPtr* ptr) {
-    Wrapper* p = internal::WrapperTLSGroup<DoublyBufferedData>::
-        get_or_create_tls_data(_wrapper_key);
+template <typename T, typename TLS, bool AllowSuspendBthread>
+int DoublyBufferedData<T, TLS, AllowSuspendBthread>::Read(
+    typename DoublyBufferedData<T, TLS, AllowSuspendBthread>::ScopedPtr* ptr) {
+    Wrapper* p = WrapperTLSGroup::get_or_create_tls_data(_wrapper_key);
     Wrapper* w = AddWrapper(p);
     if (BAIDU_LIKELY(w != NULL)) {
-        w->BeginRead();
-        ptr->_data = UnsafeRead();
-        ptr->_w = w;
+        if (AllowSuspendBthread) {
+            // Use reference count instead of mutex to indicate read of
+            // foreground instance, so during the read process, there is
+            // no need to lock mutex and bthread is allowed to be suspended.
+            w->BeginRead();
+            int index = -1;
+            ptr->_data = UnsafeRead(index);
+            ptr->_index = index;
+            w->AddRef(index);
+            ptr->_w = w;
+            w->BeginReadRelease();
+        } else {
+            w->BeginRead();
+            ptr->_data = UnsafeRead();
+            ptr->_w = w;
+        }
+
         return 0;
     }
     return -1;
 }
 
-template <typename T, typename TLS>
+template <typename T, typename TLS, bool AllowSuspendBthread>
 template <typename Fn>
-size_t DoublyBufferedData<T, TLS>::Modify(Fn& fn) {
+size_t DoublyBufferedData<T, TLS, AllowSuspendBthread>::Modify(Fn& fn) {
     // _modify_mutex sequences modifications. Using a separate mutex rather
     // than _wrappers_mutex is to avoid blocking threads calling
     // AddWrapper() or RemoveWrapper() too long. Most of the time, modifications
@@ -301,7 +597,12 @@ size_t DoublyBufferedData<T, TLS>::Modify(Fn& fn) {
     {
         BAIDU_SCOPED_LOCK(_wrappers_mutex);
         for (size_t i = 0; i < _wrappers.size(); ++i) {
-            _wrappers[i]->WaitReadDone();
+            // Wait read of old foreground instance done.
+            if (AllowSuspendBthread) {
+                _wrappers[i]->WaitReadDone(bg_index);
+            } else {
+                _wrappers[i]->WaitReadDone();
+            }
         }
     }
 
@@ -310,40 +611,40 @@ size_t DoublyBufferedData<T, TLS>::Modify(Fn& fn) {
     return ret2;
 }
 
-template <typename T, typename TLS>
+template <typename T, typename TLS, bool AllowSuspendBthread>
 template <typename Fn, typename Arg1>
-size_t DoublyBufferedData<T, TLS>::Modify(Fn& fn, const Arg1& arg1) {
-    internal::Closure1<T, Fn, Arg1> c(fn, arg1);
+size_t DoublyBufferedData<T, TLS, AllowSuspendBthread>::Modify(Fn& fn, const Arg1& arg1) {
+    Closure1<Fn, Arg1> c(fn, arg1);
     return Modify(c);
 }
 
-template <typename T, typename TLS>
+template <typename T, typename TLS, bool AllowSuspendBthread>
 template <typename Fn, typename Arg1, typename Arg2>
-size_t DoublyBufferedData<T, TLS>::Modify(
+size_t DoublyBufferedData<T, TLS, AllowSuspendBthread>::Modify(
     Fn& fn, const Arg1& arg1, const Arg2& arg2) {
-    internal::Closure2<T, Fn, Arg1, Arg2> c(fn, arg1, arg2);
+    Closure2<Fn, Arg1, Arg2> c(fn, arg1, arg2);
     return Modify(c);
 }
 
-template <typename T, typename TLS>
+template <typename T, typename TLS, bool AllowSuspendBthread>
 template <typename Fn>
-size_t DoublyBufferedData<T, TLS>::ModifyWithForeground(Fn& fn) {
-    internal::WithFG0<T, Fn> c(fn, _data);
+size_t DoublyBufferedData<T, TLS, AllowSuspendBthread>::ModifyWithForeground(Fn& fn) {
+    WithFG0<Fn> c(fn, _data);
     return Modify(c);
 }
 
-template <typename T, typename TLS>
+template <typename T, typename TLS, bool AllowSuspendBthread>
 template <typename Fn, typename Arg1>
-size_t DoublyBufferedData<T, TLS>::ModifyWithForeground(Fn& fn, const Arg1& arg1) {
-    internal::WithFG1<T, Fn, Arg1> c(fn, _data, arg1);
+size_t DoublyBufferedData<T, TLS, AllowSuspendBthread>::ModifyWithForeground(Fn& fn, const Arg1& arg1) {
+    WithFG1<Fn, Arg1> c(fn, _data, arg1);
     return Modify(c);
 }
 
-template <typename T, typename TLS>
+template <typename T, typename TLS, bool AllowSuspendBthread>
 template <typename Fn, typename Arg1, typename Arg2>
-size_t DoublyBufferedData<T, TLS>::ModifyWithForeground(
+size_t DoublyBufferedData<T, TLS, AllowSuspendBthread>::ModifyWithForeground(
     Fn& fn, const Arg1& arg1, const Arg2& arg2) {
-    internal::WithFG2<T, Fn, Arg1, Arg2> c(fn, _data, arg1, arg2);
+    WithFG2<Fn, Arg1, Arg2> c(fn, _data, arg1, arg2);
     return Modify(c);
 }
